+++ CesiumGltf/src/FeatureIdTextureView.cpp
@@ -1,5 +1,6 @@
 #include "CesiumGltf/FeatureIdTextureView.h"
 
+#include "CesiumGltf/ExtensionKhrTextureTransform.h"
 #include "CesiumGltf/SamplerUtility.h"
 
 namespace CesiumGltf {
@@ -8,16 +9,21 @@ FeatureIdTextureView::FeatureIdTextureView() noexcept
       _texCoordSetIndex(0),
       _channels(),
       _pImage(nullptr),
-      _pSampler(nullptr) {}
+      _pSampler(nullptr),
+      _applyTextureTransform(false),
+      _textureTransform(std::nullopt) {}
 
 FeatureIdTextureView::FeatureIdTextureView(
     const Model& model,
-    const FeatureIdTexture& featureIdTexture) noexcept
+    const FeatureIdTexture& featureIdTexture,
+    bool applyKhrTextureTransform) noexcept
     : _status(FeatureIdTextureViewStatus::ErrorUninitialized),
       _texCoordSetIndex(featureIdTexture.texCoord),
       _channels(),
       _pImage(nullptr),
-      _pSampler(nullptr) {
+      _pSampler(nullptr),
+      _applyTextureTransform(applyKhrTextureTransform),
+      _textureTransform(std::nullopt) {
   int32_t textureIndex = featureIdTexture.index;
   if (textureIndex < 0 ||
       static_cast<size_t>(textureIndex) >= model.textures.size()) {
@@ -72,11 +78,23 @@ FeatureIdTextureView::FeatureIdTextureView(
   this->_channels = channels;
 
   this->_status = FeatureIdTextureViewStatus::Valid;
-}
+
+  const ExtensionKhrTextureTransform* pTextureTransform =
+      featureIdTexture.getExtension<ExtensionKhrTextureTransform>();
+  if (pTextureTransform) {
+    this->_textureTransform = KhrTextureTransform(*pTextureTransform);
+  }
+} // namespace CesiumGltf
 
 int64_t FeatureIdTextureView::getFeatureID(double u, double v) const noexcept {
   if (this->_status != FeatureIdTextureViewStatus::Valid) {
     return -1;
+  }
+
+  if (this->_applyTextureTransform && this->_textureTransform) {
+    glm::dvec2 transformedUvs = this->_textureTransform->applyTransform(u, v);
+    u = transformedUvs.x;
+    v = transformedUvs.y;
   }
 
   u = applySamplerWrapS(u, this->_pSampler->wrapS);